exercise-and-nutrition-backend/requirements.txt
-r stub/requirements.txt
DBUtils
PyMySQL
connexion<3.0.0
urllib3==1.26.6
flask-cors